% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.r
\name{assert_fields}
\alias{assert_fields}
\title{Assert Fields}
\usage{
assert_fields(df, required_fields)
}
\arguments{
\item{df}{A data frame.}

\item{required_fields}{A character vector of required fields. If
missing, will default to
\code{fields(c("path", "event", "date", "value"))}.}
}
\description{
Assert that required fields exist in data.
}
