% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframeutils.r
\name{assign_baseyear}
\alias{assign_baseyear}
\title{Add Base Years}
\usage{
assign_baseyear(df, event_years)
}
\arguments{
\item{df}{A dataframe.}

\item{event_years}{A dataframe containing at minimum the fields
"event" and "baseyear". Fields "alt", "run", and
"wateryear" will also be used to attach baseyear information if
available.}
}
\value{
The dataframe \code{df}, with additional fields "baseyear".
}
\description{
Add base year information to a dataset.
}
