% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icf.r
\name{assign_icf}
\alias{assign_icf}
\title{Assign ICF Periods}
\usage{
assign_icf(df, icf_periods)
}
\arguments{
\item{df}{A dataframe.}

\item{icf_periods}{A data frame of ICF classes. At minimum, the
fields "event", "date", and "icf" are required. Additional
fields "alt" and "run" are also used to assign categories to
\code{df} if they are present in \code{icf_periods}.}
}
\value{
The dataset, with additional field "icf".
}
\description{
Label dates based on an existing ICF classification.
}
