% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframeutils.r
\name{assign_probabilities}
\alias{assign_probabilities}
\title{Assign Probabilities}
\usage{
assign_probabilities(df, event_weights, cumulative = TRUE)
}
\arguments{
\item{df}{A dataframe.}

\item{event_weights}{A dataframe with fields "event" and "wypercentile".
If present, fields "alt" and "run" will also be used when joining
probabilities. If \code{NULL}, equal weight is assigned to each event.}

\item{cumulative}{If \code{TRUE}, return cumulative probabilities.}
}
\description{
Assign incremental probabilities to events.
}
\keyword{internal}
