% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframeutils.r
\name{assign_wycategory}
\alias{assign_wycategory}
\title{Assign Water Year Category}
\usage{
assign_wycategory(df, wy_categories)
}
\arguments{
\item{df}{A dataframe.}

\item{wy_categories}{A data frame of categories. At minimum, the fields
"wateryear" and "wycategory" are required. Additional fields
"alt", "run" and "event" are also used to assign categories to
\code{df} if they are present in \code{wy_categories}.}
}
\value{
The dataset, with additional field "wycategory".
}
\description{
Label water years based on an existing classification.
}
