% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icf.r
\name{classify_icf}
\alias{classify_icf}
\title{Classify ICF Dates}
\usage{
classify_icf(df)
}
\arguments{
\item{df}{A numeric vector, typically the numeric representation of
the ICF date.}
}
\value{
The dataset, with additional field "icf".
}
\description{
Classify an ICF Dates.
}
