% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframeutils.r
\name{classify_wateryear}
\alias{classify_wateryear}
\title{Classify Water Years}
\usage{
classify_wateryear(df, cutoffs)
}
\arguments{
\item{df}{A dataframe.}

\item{cutoffs}{A named vector of cutoffs defining
the upper ends of the water year categories. The cutoffs and labels
are passed to \code{\link[base:cut]{base::cut()}}, with \code{Inf} appended to cutoffs to
complete the definition of the highest class. The lowest class
cutoff should typically be specified as \code{0} or \code{-Inf} to avoid
\code{NA} values being returned.}
}
\value{
The dataset, with additional field "wycategory".
}
\description{
Label water years based on cutoffs.
}
