% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorutils.r
\name{color_to_hex}
\alias{color_to_hex}
\title{Named Color to Hexadecimal}
\usage{
color_to_hex(color, alpha = 1)
}
\arguments{
\item{color}{A vector of named colors.}

\item{alpha}{A vector of alpha values. Standard vector recycling
rules apply if length of \code{alpha} does not match \code{color}.}
}
\value{
A vector of hexadecimal strings.
}
\description{
Convert a vector of named colors to a vector of hexidecimal strings.
}
\keyword{internal}
