% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorutils.r
\name{color_to_rgba}
\alias{color_to_rgba}
\title{Named Color to RGB Matrix}
\usage{
color_to_rgba(color, alpha = 1)
}
\arguments{
\item{color}{A vector of named colors.}

\item{alpha}{A vector of alpha values. Standard vector recycling
rules apply if length of \code{alpha} does not match \code{color}.}
}
\value{
A matrix with four rows R, B, G, A and column.
}
\description{
Convert a vector of named colors to an RGBA matrix. This is
basically a wrapper for \code{\link[grDevices:col2rgb]{grDevices::col2rgb()}} that supports
specification of numeric \code{alpha} values.
}
\keyword{internal}
