% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathutils.r
\name{drop_path_part}
\alias{drop_path_part}
\title{Drop Path Part}
\usage{
drop_path_part(path, part = c("A", "B", "C", "D", "E", "F"))
}
\arguments{
\item{path}{A vector of paths.}

\item{part}{A path part, i.e., one of \code{"A"}, \code{"B"}, \code{"C"}, \code{"D"},
\code{"E"}, \code{"F"}.}
}
\value{
A vector of paths.
}
\description{
Drop specific parts from of a path.
}
\examples{
replace_path_part("/a/b/c/d/e/f/", A = "foo", C = "", F = ".*")

}
