% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetimeutils.r
\name{end_of_month}
\alias{end_of_month}
\title{End Of Month}
\usage{
end_of_month(datetime, keep_times = FALSE)
}
\arguments{
\item{datetime}{A vector of datetimes.}

\item{keep_times}{If \code{TRUE}, keep the HMS component of the input
datetimes. If \code{FALSE}, return "23:59:59" as the HMS component.}
}
\value{
A datetime vector
}
\description{
Retrieve end of month datetimes for a set of datetimes.
}
\keyword{internal}
