% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframeutils.r
\name{extract_baseyear}
\alias{extract_baseyear}
\title{Extract Base Years}
\usage{
extract_baseyear(df, base_year)
}
\arguments{
\item{df}{A dataframe.}

\item{base_year}{An integer vector of base years.}
}
\value{
A subset of \code{df}.
}
\description{
Extract base years from a dataset.
}
