% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringutils.r
\name{extract_between}
\alias{extract_between}
\title{Extract Text Between}
\usage{
extract_between(string, before, after)
}
\arguments{
\item{string}{A character vector.}

\item{before}{A regex string preceding the text to extract.
capture groups \verb{()} are not currently supported.}

\item{after}{A regex string following the text to extract.
capture groups \verb{()} are not currently supported.}
}
\value{
A character vector of extracted strings. Unmatched strings
are returned unchanged.
}
\description{
Extract text between two strings.
}
\keyword{internal}
