% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icf.r
\name{extract_icf}
\alias{extract_icf}
\title{Extract ICF Period}
\usage{
extract_icf(df, icf_period)
}
\arguments{
\item{df}{A dataframe.}

\item{icf_period}{A vector of ICF period labels.}
}
\value{
A subset of \code{df}.
}
\description{
Extract ICF Periods from a dataset.
}
