% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframeutils.r
\name{extract_peaks}
\alias{extract_peaks}
\title{Extract Peaks}
\usage{
extract_peaks(df, by = fields("event"))
}
\arguments{
\item{df}{A dataframe.}

\item{by}{The field to group by. Default is \code{"event"}.}
}
\value{
A subset of \code{df} containing peak records.
}
\description{
Extract the peak values from a dataset.
}
