% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframeutils.r
\name{extract_window}
\alias{extract_window}
\title{Extract Window}
\usage{
extract_window(df, time_window)
}
\arguments{
\item{df}{A dataframe.}

\item{time_window}{A time window specification. Can be one of
\code{"annual"}, \code{"spring"}, \code{"winter"}, or a user-defined window
in the format \code{"jan"}, \code{"jan-mar"}, \code{"01jan-31mar"}, etc.
Time stamps such as \code{"30Apr 00:01"} or \code{"01Apr 00:01-30Apr 24:00"}
are also supported. If times are ommitted the window is assumed to
be inclusive, i.e., \code{"01jan-31mar"} will be interpreted as
\code{"01jan 00:00-31mar 24:00"}.}
}
\value{
A subset of \code{df} containing the specified time window for
from each year, with additional "period" field.
}
\description{
Extract time windows from a dataset.
}
\details{
Note that leap days will be implicitly removed when the
reference year is not a leap year, even if \code{rm.leap = FALSE}. In
the edge case of \code{time_window = "feb-jan"}, leap days are
removed according to whether the year preceding \code{reference_year}
is a leap year, since February will be associated with the
preceding year rather than the reference year.
}
