% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframeutils.r
\name{extract_wycategory}
\alias{extract_wycategory}
\title{Extract Water Year Category}
\usage{
extract_wycategory(df, wy_category)
}
\arguments{
\item{df}{A dataframe.}

\item{wy_category}{A vector of water year categories.}
}
\value{
A subset of \code{df}.
}
\description{
Extract water year categories from a dataset.
}
