% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.r
\name{fields}
\alias{fields}
\title{Wat Plotter Fields}
\usage{
fields(...)
}
\arguments{
\item{...}{A vector of field codes, any of: "alt", "run", "path",
"event", "date", "value", "refdate", "wateryear", "wycategory",
"wypercentile", "baseyear".}
}
\value{
A vector of field names.
}
\description{
Standard Fields in WAT Plotter data frames.
}
