% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicutils.r
\name{first_max}
\alias{first_max}
\title{First Max}
\usage{
first_max(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A logical vector of same length as \code{x}.
}
\description{
Identify the first occurence of the maximum value in a vector.
}
\keyword{internal}
