% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathutils.r
\name{form_path}
\alias{form_path}
\title{Form Path}
\usage{
form_path(A = "", B = "", C = "", D = "", E = "", F = "", parts = list())
}
\arguments{
\item{A}{A vector of "A" path parts.}

\item{B}{A vector of "B" path parts.}

\item{C}{A vector of "C" path parts.}

\item{D}{A vector of "D" path parts.}

\item{E}{A vector of "E" path parts.}

\item{F}{A vector of "F" path parts.}

\item{parts}{A named list of path parts.}
}
\value{
A vector of paths.
}
\description{
Form path from parts.
}
