% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathutils.r
\name{get_path_part}
\alias{get_path_part}
\title{Get Path Part}
\usage{
get_path_part(path, part = c("A", "B", "C", "D", "E", "F"))
}
\arguments{
\item{path}{A vector of paths.}

\item{part}{A path part, i.e., one of \code{"A"}, \code{"B"}, \code{"C"}, \code{"D"},
\code{"E"}, \code{"F"}.}
}
\value{
A vector of path parts.
}
\description{
Extract a specific part from a path.
}
