% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetimeutils.r
\name{get_reference_year}
\alias{get_reference_year}
\title{Reference Year}
\usage{
get_reference_year(reftime)
}
\arguments{
\item{reftime}{A vector of reference datetimes, e.g., output of
\code{\link[=reference_date]{reference_date()}}.}
}
\value{
A vector of integer reference years.
}
\description{
Get the reference water year from a timestamp.
}
\keyword{internal}
