% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wycategory-tables.r
\name{list_standard_cutoffs}
\alias{list_standard_cutoffs}
\title{Standard Water Year Category Tables}
\usage{
list_standard_cutoffs(name = NULL, project = NULL, interval = NULL)
}
\arguments{
\item{name}{A table name to search for.}

\item{project}{A project name to search for.}

\item{interval}{An interval to search for.}
}
\value{
A tibble.
}
\description{
List the standard water year category tables.
}
