% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicutils.r
\name{notin-infix}
\alias{notin-infix}
\alias{\%notin\%}
\title{Not In}
\usage{
x \%notin\% table
}
\arguments{
\item{x}{vector or \code{NULL}: the values to be matched.
    \link[base]{Long vectors} are supported.}

\item{table}{vector or \code{NULL}: the values to be matched against.
    \link[base]{Long vectors} are not supported.}
}
\value{
A logical vector, indicating if a match was not located
for each element of x: thus the values are \code{TRUE} or \code{FALSE}
and never \code{NA}.
}
\description{
Infix operator equivalent to \code{!(x \%in\% table)}.
}
\seealso{
\code{\link[base:match]{base::match()}}
}
\keyword{internal}
