% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathutils.r
\name{path_to_parts}
\alias{path_to_parts}
\title{Path To Parts}
\usage{
path_to_parts(path)
}
\arguments{
\item{path}{A vector of paths.}
}
\value{
A dataframe with fields "A", "B", "C", "D", "E", "F"
containing the path parts.
}
\description{
Split a vector of paths into a dataframe of parts.
}
