% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetimeutils.r
\name{reference_date}
\alias{reference_date}
\title{Reference Date}
\usage{
reference_date(datetime, reference_year = 3001L)
}
\arguments{
\item{datetime}{A vector of dates or datetimes.}

\item{reference_year}{The reference year to use when returning
datetimes. Ignored when \code{as_date = FALSE}.}
}
\value{
A vector of \code{POSIXct} reference timestamps
}
\description{
Return datetimes with a fixed water year. Useful for overplotting
timeseries.
}
\keyword{internal}
