% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframeutils.r
\name{rename_fields}
\alias{rename_fields}
\title{Rename Fields}
\usage{
rename_fields(df, old_names, new_names)
}
\arguments{
\item{df}{A dataframe.}

\item{old_names}{A vector of names that may be in \code{df}.}

\item{new_names}{A vector of names to replace those in \code{old_names}.}
}
\value{
The dataframe, potentially with new fields names.
}
\description{
Rename fields in a data frame.
}
