% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathutils.r
\name{replace_path_part}
\alias{replace_path_part}
\title{Replace Path Part}
\usage{
replace_path_part(path, ...)
}
\arguments{
\item{path}{A vector of paths.}

\item{...}{Path parts to replace.}
}
\value{
A vector of paths.
}
\description{
Replace specific parts from of a path.
}
\examples{
replace_path_part("/a/b/c/d/e/f/", A = "foo", C = "", F = ".*")

}
