% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorutils.r
\name{rgba_to_hex}
\alias{rgba_to_hex}
\title{RGBA Matrix to Hexadecimal}
\usage{
rgba_to_hex(rgba)
}
\arguments{
\item{rgba}{A matrix with four rows for Red, Green, Blue, and
Alpha channels, i.e., output of \code{\link[=color_to_rgba]{color_to_rgba()}}.}
}
\value{
A vector of hexadecimal strings.
}
\description{
Convert an RGBA matrix to a vector of hexidecimal strings.
}
\keyword{internal}
