% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframeutils.r
\name{rolling_window}
\alias{rolling_window}
\title{Rolling Window}
\usage{
rolling_window(df, window_size, fun = mean, ..., complete = TRUE)
}
\arguments{
\item{df}{A dataframe.}

\item{window_size}{The rolling window size. Assumes a centered
rolling window.}

\item{fun}{The function to apply to the rolling window.}

\item{...}{Additional parameters to pass to \code{fun}.}

\item{complete}{If \code{TRUE}, evaluate \code{fun} on complete windows only.
Incomplete sections will be filled with \code{NA}.}
}
\description{
Apply a rolling window function to a dataset.
}
