% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wycategory-tables.r
\name{standard_cutoffs}
\alias{standard_cutoffs}
\title{Get Standard Water Year Cutoffs}
\usage{
standard_cutoffs(name)
}
\arguments{
\item{name}{A table name to search for.}
}
\value{
A named numeric vector of cutoffs, with additional
attributes \code{"project"}, \code{"interval"}, and  \code{"date"}.
}
\description{
Get the cutoffs from a standard water year category table.
}
