% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframeutils.r
\name{standardize_dataset}
\alias{standardize_dataset}
\title{Standardize Dataset}
\usage{
standardize_dataset(df, rm.leap = TRUE, reference_year = 3001L)
}
\arguments{
\item{df}{A dataframe.}

\item{rm.leap}{If \code{TRUE}, remove leap days.}

\item{reference_year}{The reference year to use when returning
datetimes. Ignored when \code{as_date = FALSE}.}
}
\value{
The dataframe, with new fields.
}
\description{
Standardize a WAT Plotter dataset as follows:
}
\details{
\enumerate{
\item Assert required fields are present.
\item Add water year field (integer).
\item Add reference date field (Date).
}

timestamps in a dataframe to a reference water year.

Note that leap days will be converted to \code{NA} if
\code{reference_year} is not a leap year.
}
