% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringutils.r
\name{string-truncation}
\alias{string-truncation}
\alias{left}
\alias{right}
\title{String Left/Right Extraction}
\usage{
left(string, width)

right(string, width)
}
\arguments{
\item{string}{Input vector. Either a character vector, or something
coercible to one.}

\item{width}{Maximum width of string.}
}
\value{
A left- or right-truncated string.
}
\description{
Extract The leftmost/rightmost characters from a string.
}
\keyword{internal}
