% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetimeutils.r
\name{wateryear}
\alias{wateryear}
\title{Water Year}
\usage{
wateryear(datetime, reference_year = 3001L)
}
\arguments{
\item{datetime}{A vector of dates or datetimes.}
}
\value{
A vector of integer water years.
}
\description{
Identify the water year of the provided datetimes.
}
\keyword{internal}
