% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\docType{package}
\name{watplotter.core-package}
\alias{watplotter.core}
\alias{watplotter.core-package}
\title{watplotter.core: WAT Plotter Core Utilities}
\description{
Core functions and expectations for watplotter packages. This package is designed as a common dependency and is not for direct use.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://nwd-wmshare.nwd.usace.army.mil:3000/git/DataMgmt/watplotter.core}
  \item Report bugs at \url{http://nwd-wmshare.nwd.usace.army.mil:3000/git/DataMgmt/watplotter.core/issues}
}

}
\author{
\strong{Maintainer}: Michael Koohafkan \email{Michael.C.Koohafkan@usace.army.mil}

Other contributors:
\itemize{
  \item Trey Crouch \email{Trey.D.Crouch@usace.army.mil} [contributor]
}

}
\keyword{internal}
