% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetimeutils.r
\name{window_to_range}
\alias{window_to_range}
\title{Window to Range}
\usage{
window_to_range(time_window, reference_year = 3001L)
}
\arguments{
\item{time_window}{A time window specification. Can be one of
\code{"annual"}, \code{"spring"}, \code{"winter"}, or a user-defined window
in the format \code{"jan"}, \code{"jan-mar"}, \code{"01jan-31mar"}, etc.
Time stamps such as \code{"30Apr 00:01"} or \code{"01Apr 00:01-30Apr 24:00"}
are also supported. If times are ommitted the window is assumed to
be inclusive, i.e., \code{"01jan-31mar"} will be interpreted as
\code{"01jan 00:00-31mar 24:00"}.}
}
\value{
A two-element date vector.
}
\description{
Convert a time window to a date range using a reference year.
}
\keyword{internal}
